requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==23.2.0
    # via
    #   glom
    #   jsonschema
    #   referencing
    #   semgrep
boltons==21.0.0
    # via
    #   face
    #   glom
    #   semgrep
bracex==2.4
    # via wcmatch
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   click-option-group
    #   semgrep
click-option-group==0.5.6
    # via semgrep
colorama==0.4.6
    # via
    #   click
    #   semgrep
defusedxml==0.7.1
    # via semgrep
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
exceptiongroup==1.2.2
    # via semgrep
face==22.0.0
    # via glom
fire==0.6.0
    # via -r requirements.in
future-fstrings==1.2.0
    # via pyecore
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via -r requirements.in
glom==22.1.0
    # via semgrep
googleapis-common-protos==1.63.2
    # via opentelemetry-exporter-otlp-proto-http
graphviz==0.20.3
    # via -r requirements.in
idna==3.7
    # via requests
importlib-metadata==7.1.0
    # via opentelemetry-api
jsonschema==4.23.0
    # via semgrep
jsonschema-specifications==2023.12.1
    # via jsonschema
lark==1.1.9
    # via python-hcl2
lxml==5.2.2
    # via pyecore
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
neo4j==5.22.0
    # via -r requirements.in
opentelemetry-api==1.25.0
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   semgrep
opentelemetry-exporter-otlp-proto-common==1.25.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.25.0
    # via semgrep
opentelemetry-instrumentation==0.46b0
    # via opentelemetry-instrumentation-requests
opentelemetry-instrumentation-requests==0.46b0
    # via semgrep
opentelemetry-proto==1.25.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.25.0
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   semgrep
opentelemetry-semantic-conventions==0.46b0
    # via
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-sdk
opentelemetry-util-http==0.46b0
    # via opentelemetry-instrumentation-requests
ordered-set==4.1.0
    # via pyecore
packaging==24.1
    # via semgrep
peewee==3.17.6
    # via semgrep
protobuf==4.25.3
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydal==20240713.1
    # via pytm
pydot==3.0.1
    # via -r requirements.in
pyecore==0.15.1
    # via -r requirements.in
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via pydot
python-hcl2==4.3.4
    # via -r requirements.in
pytm==1.3.1
    # via -r requirements.in
pytz==2024.1
    # via neo4j
pyyaml==6.0.1
    # via -r requirements.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   semgrep
restrictedpython==7.1
    # via pyecore
rich==13.7.1
    # via semgrep
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.17.40
    # via semgrep
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
semgrep==1.80.0
    # via -r requirements.in
six==1.16.0
    # via fire
smmap==5.0.1
    # via gitdb
termcolor==2.4.0
    # via fire
tomli==2.0.1
    # via semgrep
typing-extensions==4.12.2
    # via
    #   opentelemetry-sdk
    #   semgrep
urllib3==2.2.2
    # via
    #   requests
    #   semgrep
wcmatch==8.5.2
    # via semgrep
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.19.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
